fix: clear the userspace clippy backlog (172 findings across ~40 crates)
`cargo xtask check` linted 29 of 133 userspace crates, so the other 104
accumulated a backlog nobody could see. This clears it, ahead of the
companion change that puts them all under the gate.

Mostly mechanical — `collapsible_if`, `is_multiple_of`, `contains` vs
`iter().any()`, `manual_clamp`, `needless_range_loop`, needless deref,
doc-list indentation. Four things were worth more than the lint:

- **A DMA use-after-free avoided.** xhci's `SlotContext::output_ctx` reads
  as dead, but `SlotContext::new` allocates that DmaBuffer and writes its
  IOVA into the DCBAA, after which only the xHC touches the memory.
  Deleting the field would have dropped the buffer while the controller's
  DCBAA entry still pointed at it — and it would have compiled. Kept, with
  the reason written down.

- **`term/src/display.rs` had five findings host-target clippy cannot
  reach**, because the module sits behind the `os-binary` feature. That is
  the coverage gap in miniature: the file this phase edited most was
  structurally invisible to the lint.

- **A dead branch in the dwindle layout**, left alone deliberately: both
  arms of `if self.rotate { depth % 2 == 0 } else { depth % 2 == 0 }` are
  identical, and the comment claims a rotation this expression does not
  perform. The spiral does rotate — via the carve swap further down — so
  the layout is not broken and nothing pins the intended behaviour. Fixing
  it means choosing new tiling geometry, which is not a lint decision.

- **A `heapless` truncation in the compositor's flip path** (compose.rs):
  `extend_from_slice` into a capacity-2 vec is all-or-nothing and its
  `Err(())` was discarded, so the fold is a silent no-op in the common
  case. Made explicit and documented rather than silently widened —
  changing what the VBE flip backend composes needs a screenshot gate, not
  a lint pass.

56 `#[allow]`s were added where the lint is genuinely wrong for the
situation — blit primitives whose signature mirrors a framebuffer rect,
wire-ABI constants that are never read because the protocol detects
absence by label, hardware struct fields the device reads by DMA. Each
carries its reason inline; none is a blanket suppression.

Also noted for follow-up, not fixed here: `display_server`, `usb_hid` and
`usbhub` have `#[cfg(test)] mod tests` blocks that have never compiled —
their `#[panic_handler]` lacks a `#[cfg(not(test))]` gate, so
`cargo test -p display_server` dies on a duplicate `panic_impl` lang item.
Roughly 100 already-written unit tests do not run, which is why so much of
that crate reads as dead code.

+++ userspace/async-rt/src/io.rs
@@ -43,12 +43,11 @@ impl AsyncFd {
 fn check_fd_readable(fd: i32) -> bool {
     let mut pfd = libc::pollfd {
         fd,
-        events: libc::POLLIN as i16,
+        events: libc::POLLIN,
         revents: 0,
     };
     let ret = unsafe { libc::poll(&mut pfd, 1, 0) };
-    ret > 0
-        && (pfd.revents & (libc::POLLIN as i16 | libc::POLLHUP as i16 | libc::POLLERR as i16)) != 0
+    ret > 0 && (pfd.revents & (libc::POLLIN | libc::POLLHUP | libc::POLLERR)) != 0
 }
 
 #[cfg(not(feature = "std"))]
@@ -67,11 +66,11 @@ fn check_fd_readable(fd: i32) -> bool {
 fn check_fd_writable(fd: i32) -> bool {
     let mut pfd = libc::pollfd {
         fd,
-        events: libc::POLLOUT as i16,
+        events: libc::POLLOUT,
         revents: 0,
     };
     let ret = unsafe { libc::poll(&mut pfd, 1, 0) };
-    ret > 0 && (pfd.revents & libc::POLLOUT as i16) != 0
+    ret > 0 && (pfd.revents & libc::POLLOUT) != 0
 }
 
 #[cfg(not(feature = "std"))]
